+++ Templates/scripts/shared/switch-branch.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-// **Version:** 0.23.3
+// **Version:** 0.23.4
 /**
  * switch-branch.js
  *